+++ src/app/features/event/event.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { EventRoutingModule } from './event-routing.module';
+import { EventResolverComponent } from './event-resolver/event-resolver.component';
+
+
+@NgModule({
+  declarations: [
+    EventResolverComponent
+  ],
+  imports: [
+    CommonModule,
+    EventRoutingModule
+  ]
+})
+export class EventModule { }