+++ changelog.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `feature` Add event-panel for autor+ users in admin dashboard.
 - `feature` Add event-mode logo in event-admin view.
+- `feature` Add url-sanitized field for events, and evento/url view.
 
 ### Update